requirements.txt
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.7
django-autocomplete-light==3.2.10
django-debug-toolbar==1.8
django-toolbelt==0.0.1
djangorestframework==3.7.3
gunicorn==19.7.1
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
six==1.11.0
sqlparse==0.2.4
static3==0.7.0
whitenoise==3.3.1
